# spec/openapi.yaml
openapi: 3.0.0
servers:
# Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/ProtocolOne/Authentication/1.0.0
  - url: 'https://auth-one.protocol.one'
info:
  description: The Authentication API enables you to manage all aspects of user identity when you use Auth-One. It offers endpoints so your users can log in, sign up, log out, access APIs, and more.
  version: 1.0.0
  title: ProtocolOne Authentication API
  termsOfService: ''
  contact:
    email: vadim.sabirov@protocol.one
    name: Vadim Sabirov
    url: 'https://protocol.one'
...
paths:
...
  /dbconnections/change_password:
    post:
      tags:
      - Change Password
      summary: 'Send verification code for the change password'
      description: 'Given a user''s `email` address and a `connection`, Auth-One will send a verification code to the email address.'
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of the database connection configured to your client'
        - name: email
          in: query
          required: true
          schema:
            type: string
          description: 'The user''s email address'
      operationId: changePasswordStart
      responses:
        '200':
          description: Return One-Time Token for verification method
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OneTimeToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /dbconnections/change_password/verify:
    post:
      tags:
      - Change Password
      summary: 'Verify change password code'
      description: ''
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of the database connection configured to your client'
        - name: verification_code
          in: query
          required: true
          schema:
            type: string
          description: 'Verification code from email'
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: 'One-Time Token'
        - name: password
          in: query
          required: true
          schema:
            type: string
          description: 'The new password'
      operationId: changePasswordVerify
      responses:
        '200':
          description: 'User Access Token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
...
  /login:
    post:
      tags:
      - Login
      summary: 'Authenticate by login and password'
      description: 'Use this endpoint to authenticate a user with database connection.'
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of the connection configured to your application'
        - name: email
          in: query
          required: true
          schema:
            type: string
          description: 'The user''s email address'
        - name: password
          in: query
          required: true
          schema:
            type: string
          description: 'The user''s desired password'
        - name: captcha
          in: query
          required: false
          schema:
            type: string
          description: 'Captcha'
      operationId: login
      responses:
        '200':
          description: 'User Access Token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'MFA required. The `error` will be the value `mfa_required` and the `error_message` will be contains token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /authorize:
    get:
      tags:
      - Login
      summary: 'Authenticate by social provider'
      description: 'Use this endpoint to authenticate a user with a social. It will return a 302 redirect to the social provider specified in connection. The `redirect_uri` value must be specified as a valid callback URL under your Application''s Settings.<br><br>You can use a special endpoint for `redirect_uri` if you want to get the Access Token of a user via JS PostMessage or WebSocket. See more on <a href="#/Login/authorizeResult">authorizeResult</a> page'
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of a social identity provider configured to your application, for example `google-oauth2` or `facebook`'
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
          description: 'The URL to which will redirect the browser after authorization has been granted by the user'
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: 'An opaque value the applications adds to the initial request that the authorization server includes when redirecting the back to the application. This value must be used by the application to prevent CSRF attacks.'
        - name: ADDITIONAL_PARAMETERS
          in: query
          required: false
          schema:
            type: string
          description: 'Append any additional parameter to the end of your request, and it will be sent to the provider. For example, access_type=offline (for Google Refresh Tokens) , display=popup (for Windows Live popup mode)'
      operationId: authorize
      responses:
        '302':
          description: >-
            Redirect to destination URI with One-Time Token. Adds to the request
            the parameter `auth_one_ott` with the value of a one-time token.
            Allows it to receive a one-time user Access Token value. For example:
            `https://YOUR_DOMAIN/page?auth_one_ott=zcv987z...987dre987zxc`
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /authorize/result:
    get:
      tags:
      - Login
      summary: 'Send social authorize result'
      description: 'Send social authorize result via JS PostMessage or WebSocket protocol.<br><br>The result will be sent as a JSON.stringify object with the following structure: `{"event":"EVENT_NAME", "token":{}}` where the name of the event can be `oauthCompletedSuccessfully` or `oauthCompletedWithError`, and the token upon successful authentication contains a JWTToken object.'
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of a social identity provider configured to your application, for example `google-oauth2` or `facebook`'
        - name: auth_one_ott
          in: query
          required: true
          schema:
            type: string
          description: 'One-Time Token'
        - name: wsUrl
          in: query
          required: false
          schema:
            type: string
          description: 'If you want to receive an answer via WebSocket, set the value of this parameter as the address of connection to your WebSocket server, for example: `wss://127.0.0.1:123`. The parameter must be encoded with the encodeURIComponent, for example `wss%3A%2F%2F127.0.0.1%3A123`'
      responses:
        '200':
          description: Bad Request
          content:
            plain/text:
              schema:
                type: string
                example: pong
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: authorizeResult
      security: []
...
  /mfa/challenge:
    post:
      tags:
      - Multi-Factor Authentication
      summary: 'Set a MFA challenge types'
      description: 'Request a challenge based on the challenge types supported by the application and user.'
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of the connection configured to your application'
        - name: mfa_token
          in: query
          required: true
          schema:
            type: string
          description: 'MFA token from authenticate response'
        - name: challenge_type
          in: query
          required: false
          schema:
            type: integer
          description: 'Set `otp` for use one-time password or set `oob` for use out-of-band message (sms). If the parameter is not set, then priority is given to the type of `otp`, if it is connected.'
      operationId: mfaChallenge
      responses:
        '200':
          description: 'Get user Access Token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /mfa/verify:
    post:
      tags:
      - Multi-Factor Authentication
      summary: 'Authenticate a user with MFA'
      description: 'Use this endpoint to authenticate a user with MFA.'
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of the connection configured to your application'
        - name: mfa_token
          in: query
          required: true
          schema:
            type: string
          description: 'MFA token from authenticate response'
        - name: code
          in: query
          required: false
          schema:
            type: integer
          description: 'OTP, OOB or Recovery code'
      operationId: mfaVerify
      responses:
        '200':
          description: 'Get user Access Token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /mfa/add:
    post:
      tags:
      - Multi-Factor Authentication
      summary: 'Add authenticator to the user'
      description: ''
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: 'The name of the connection configured to your application'
        - name: authenticator_types
          in: query
          required: true
          schema:
            type: string
          description: 'Set `otp` for use one-time password or set `oob` for use out-of-band message (sms)'
        - name: oob_channel
          in: query
          required: false
          schema:
            type: string
          description: 'The type of OOB channels supported by the client. An array with values `auth1` or `sms`. Required if authenticator_types include `oob`'
        - name: phone_number
          in: query
          required: false
          schema:
            type: string
          description: 'The phone number to use for SMS. Required if `oob_channel` includes `sms`'
      operationId: mfaAdd
      responses:
        '200':
          description: 'Get added MFA settings'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaAuthenticator'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: bad_request
          description: 'Error code (for example: `mfa_request` for the MFA scenario or `password` for form field)'
        error_message:
          type: string
          example: The connection was not found
          description: 'Text of error or value for specific property (e.g. MFA one-time token value)'
...
    JWTToken:
      type: object
      properties:
        access_token:
          type: string
          example: oDnPKc1a...vWE36Lh
          description: JWT token
        expires_in:
          type: integer
          format: int32
          example: 3600
          description: UNIX timestamp
        id_token:
          type: string
          description: Refresh token
          example: oDnPKc1a...vWE36Lh
...
    OneTimeToken:
      type: object
      description: 'One-Time Token'
      properties:
        token:
          type: string
          example: oDnPKc1a...vWE36Lh
    MfaAuthenticator:
      type: object
      properties:
        secret:
          type: string
          example: oDnPKc1a...vWE36Lh
          description: Use for OTP authentificator
        oob_channel:
          type: string
          example: sms
          description: Use for OBB authentificator. Must be `sms` or `auth1`
        barcode_uri:
          type: string
          example: otpauth://...
        authenticator_type:
          type: string
          description: '`otp` or `oob`'
          example: otp
        recovery_codes:
          type: array
          items:
            type: string
          example: ["ABCDEFGDRFK75ABYR7PH8TJA"]
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: X-AUTH-ONE-TOKEN-{CLIENT_ID}
    BearerAuth:
      type: http
      scheme: Bearer
      bearerFormat: JWT
...
security:
  - CookieAuth: []
  - BearerAuth: []